Unify docs and release notes design
