+++ docs/src/neo/design/components/IconCard.tsx
@@ -11,7 +11,9 @@ export function IconCard({ icon }: { icon: ILibraryIcon }) {
       className="relative flex flex-col items-center justify-center gap-3 rounded-sm border border-zinc-950/10 p-4 transition-all focus-within:bg-violet-100 focus-within:text-violet-700 hover:border-violet-700/15 hover:bg-violet-100 hover:text-violet-700 hover:shadow-sm"
     >
       <Amicon icon={icon.icon} className="mt-1 text-3xl" />
-      <span className="max-w-full truncate font-mono text-xs text-nowrap text-zinc-700">{icon.slug}</span>
+      <span className="max-w-full truncate font-mono text-xs text-nowrap text-zinc-700">
+        {icon.slug}
+      </span>
     </NavLink>
   );
 }
