+++ docs/src/style.css
@@ -96,6 +96,12 @@ pre {
   }
 }
 
+.neo-doc-page {
+  & > h1:first-child {
+    display: none;
+  }
+}
+
 .neo-docs {
   :is(ul, ol) {
     margin-bottom: 0.75rem;
